Refactored NumPy module to get valid import sequence
