+++ numba/core/datamodel/models.py
@@ -319,7 +319,6 @@ class EnumModel(ProxyModel):
 @register_default(types.ExternalFunction)
 @register_default(types.EnumClass)
 @register_default(types.IntEnumClass)
-@register_default(types.NumberClass)
 @register_default(types.TypeRef)
 @register_default(types.NamedTupleClass)
 @register_default(types.RecursiveCall)
