+++ numba/cpython/builtins.py
@@ -354,26 +354,6 @@ def complex_impl(context, builder, sig, args):
     res = cmplx._getvalue()
     return impl_ret_untracked(context, builder, sig.return_type, res)
 
-
-@lower_builtin(types.NumberClass, types.Any)
-def number_constructor(context, builder, sig, args):
-    """
-    Call a number class, e.g. np.int32(...)
-    """
-    if isinstance(sig.return_type, types.Array):
-        # Array constructor
-        dt = sig.return_type.dtype
-        def foo(*arg_hack):
-            return np.array(arg_hack, dtype=dt)
-        res = context.compile_internal(builder, foo, sig, args)
-        return impl_ret_untracked(context, builder, sig.return_type, res)
-    else:
-        # Scalar constructor
-        [val] = args
-        [valty] = sig.args
-        return context.cast(builder, val, valty, sig.return_type)
-
-
 #-------------------------------------------------------------------------------
 # Constants
