+++ numba/np/np_install.py
@@ -25,3 +25,22 @@ npy_scalar_cast = lambda self, value: getattr(np, self.name)(value)
 types.Integer.cast_python_value = npy_scalar_cast
 types.Float.cast_python_value = npy_scalar_cast
 types.Complex.cast_python_value = npy_scalar_cast
+
+
+# Install a unification method for Number types that
+# uses NumPy's type promotion rules.
+def unify_number(self, typingctx, other):
+    """
+    Unify the two number types using Numpy's rules.
+    """
+    from numba.np import numpy_support
+    if isinstance(other, types.Number):
+        # XXX: this can produce unsafe conversions,
+        # e.g. would unify {int64, uint64} to float64
+        a = numpy_support.as_dtype(self)
+        b = numpy_support.as_dtype(other)
+        sel = np.promote_types(a, b)
+        return numpy_support.from_dtype(sel)
+
+
+types.Number.unify = unify_number
