+++ numba/np/npyimpl.py
@@ -875,18 +875,3 @@ def numpy_dtype(dtype):
     else:
         raise errors.NumbaTypeError(
             'unknown dtype descriptor: {}'.format(dtype))
-
-def unify_number(self, typingctx, other):
-    """
-    Unify the two number types using Numpy's rules.
-    """
-    from numba.np import numpy_support
-    if isinstance(other, types.Number):
-        # XXX: this can produce unsafe conversions,
-        # e.g. would unify {int64, uint64} to float64
-        a = numpy_support.as_dtype(self)
-        b = numpy_support.as_dtype(other)
-        sel = np.promote_types(a, b)
-        return numpy_support.from_dtype(sel)
-
-types.Number.unify = unify_number
